Create a module for common plist methods
Previously, common methods that are going to be necessary for most of the
providers in this module were stuck inside the cfpropertylist service
provider. I wanted to break them out into their own namespace so other
provdiers could us their functionality without having to redefine it all
over the place. I'm using the PuppetX namespace per Puppet Labs
recommendations, and calling the library 'Plist' so I don't have a namespace
clash with 'CFPropertyList' (wasn't this the very problem this thing was
supposed to fix?).